requirements.txt
ccxt
pandas
pandas-ta
python-dotenv
python-telegram-bot
schedule
